tee-supplicant: create profiling files with permission 0644
Profiling files are created in /tmp by tee-supplicant with permissions
0600 (-rw-------). In other words, only the owner of the file can
access it. This is a bit annoying because tee-supplicant typically runs
under a special user account ('tee') which is not normally used to log
into the system. As a result, the files are difficult to access when
debugging.

Change permissions to 0644 (-rw-r--r--) for easier access to the data.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Sumit Garg <sumit.garg@linaro.org>