rpmb emu: fix ioctl return type
The RPMB spec defines return results of commands as a single byte
value. A result from an ioctl call for a command is stored in
'op_result' of 'struct rpmb_data_frame', which is of type 'uint16_t',
which corresponds to 'uint8_t op_result[2]' of 'struct rpmb_data_frame'
in 'optee_os', with the single byte result stored @ index 1 of the
2-element byte array since data transfer is MSB first, so we need to
convert the return results into the proper uint16_t values.

Signed-off-by: Victor Chong <victor.chong@linaro.org>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>