libteec: Avoid memcpy() when using TEEC_TempMemoryReference
When dynamic shared memory is enabled, use TEEC_RegisterSharedMemory()
instead of TEE_AllocateSharedMemory() when dealing with parameters of
type TEEC_TempMemoryReference.

In the current implementation, when dealing with such parameters,
a new temporary buffer is always allocated and then registered as
shared memory. As a consequnce, memcpy() is always required to/from
this temporary buffer based on the parameter type (input/output/inout).
If the buffer passed by user can be directly registered as shared memory,
this allocation of temporary buffer is not required. This helps in
avoiding unnecessary memcpy.

Acked-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>