libckteec: serialize: Fix crash with NULL pValue in attribute
An attribute template (type, pValue, ulValueLen) provided with
C_GetAttributeValue() may have pValue field as NULL pointer with
a zero/non-zero ulValueLen field. In these cases TEE should modify
the ulValueLen to hold the exact length of the specified attribute
for the object.

In the current code, for non-zero ulValueLen with pValue as NULL
pointer, an attempt was being made to read the ulValueLen length
data from NULL pointer resulting in segmentation fault. To fix this,
in such scenarios, pass ulValueLen as 0 to TEE.

Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Vesa Jääskeläinen <vesa.jaaskelainen@vaisala.com>
Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>