libckteec: In C_OpenSession() allow notify callback to be specified
Some libraries registers notify callbacks to Cryptoki just to be on safe side
without having intention of aborting the operation.

Possibly uses for notify callback are defined in:

PKCS #11 Cryptographic Token Interface Base Specification
Version 2.40 Plus Errata 01
5.16 Callback functions

"A typical use of a surrender callback might be to give an application user
feedback during a lengthy key pair generation operation.  Each time the
application receives a callback, it could display an additional “.” to the user."

OP-TEE do not currently have this kind of mechanism to abort key generation
thus no need for such callback.

If callee has registered for such callback and is prepared to display progress
indication or abort the progress it will just never happen. User application
receives control back upon completion of such operaton.

It also states that:

"A Cryptoki library is not required to make any surrender callbacks."

So with the change we take liberty of not making the callbacks at all.

Signed-off-by: Vesa Jääskeläinen <vesa.jaaskelainen@vaisala.com>
Acked-by: Ruchika Gupta <ruchika.gupta@linaro.org>