tee-supplicant: read_with_timeout(): fix non-blocking peeking
In case read_with_timeout() is called with *blen == 0 it indicates that
read_with_timeout() should just check how much data is available, and
set how much in *blen and return success. If there is none available it
just also returns success.

Prior to this patch read_with_timeout() was instead returning an error
in case no data was available, this is not expected by the caller so fix
this by returning success in that case instead.

Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>