tee-supplicant: read rpmb dev info from sysfs
For now tee-supplicant retrieve RPMB informations by opening mmc dev
node and read data at the corresponding offset.
That is a problem on Android device with selinux enabled because the
access to this node is restricted to kernel/init.

All the necessary informations can instead be read from sysfs:
- RPMB partition size:
/sys/class/mmc_host/mmc%u/mmc%u:0001/raw_rpmb_size_mult
- Reliable write sector count:
/sys/class/mmc_host/mmc%u/mmc%u:0001/rel_sectors

Tests done on the following board: (Android build)
https://ologicinc.com/portfolio/mediatek-pumpkin-i350/

Signed-off-by: Julien Masson <jmasson@baylibre.com>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (hikey-hikey)